requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.46.0
humanfriendly==10.0
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tifffile==2023.9.26
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
loguru==0.7.2
faiss-cpu==1.7.4
filterpy==1.4.5
lap==0.5.12
torch>=2.0.0
torchvision>=0.15.0
